chore: move API logic from forge package to MCC app
