+++ packages/financial-eligibility-journey/src/index.ts
@@ -1,6 +1,7 @@
 import { createForgePackage, createFunctionsRegistry, type ForgePackage } from '@ministryofjustice/hmpps-forge/core/authoring'
 import { eligibilityJourney } from './journey.js'
-import { type Deps, FinancialEligibilityEffectsImplementations } from './effects.js'
+import { FinancialEligibilityEffectsImplementations } from './effects.js'
+import { type Deps } from './api.js'
 
 
 /**
